refactor: Clean up MUI components and unused components along with replacing MUI icon for lucide icon. Generalize image place holder.
