+++ Frontend/components/core/NoItem.tsx
@@ -1,13 +1,12 @@
 "use client";
 
 import { Container, Stack, Text, Title } from "@mantine/core";
-import { SvgIconComponent } from "@mui/icons-material";
 import { LucideIcon } from "lucide-react";
 
 type NoItemProps = {
   title?: string;
   description: string;
-  icon?: SvgIconComponent | LucideIcon;
+  icon?: LucideIcon;
 };
 
 export const NoItem = ({
@@ -17,18 +16,10 @@ export const NoItem = ({
 }: NoItemProps) => {
   return (
     <Container size="sm">
-      <Stack
-        justify="center"
-        align="center"
-        mih='60vh'
-      >
+      <Stack justify="center" align="center" mih="60vh">
         {Icon && <Icon />}
-        <Title order={2}>
-          {title}
-        </Title>
-        <Text c='dimmed'>
-          {description}
-        </Text>
+        <Title order={2}>{title}</Title>
+        <Text c="dimmed">{description}</Text>
       </Stack>
     </Container>
   );
