--- Frontend/components/data-table/MantineDataTableImageRender.tsx
@@ -1,27 +1,0 @@
-"use client";
-
-import { GridRenderCellParams } from "@mui/x-data-grid";
-import Image from "next/image";
-
-export const ProductImageRenderer = (
-  params: GridRenderCellParams<{ imageUrls: string[]; sku: string }>
-) => {
-  return (
-    <Image
-      src={
-        params.row.imageUrls?.length > 0
-          ? params.row.imageUrls[0]
-          : "/image-not-found.jpg"
-      }
-      alt={params.row.sku}
-      width={300}
-      height={0}
-      style={{
-        width: "100%",
-        height: "100%",
-        objectFit: "fill",
-        aspectRatio: 1 / 1,
-      }}
-    />
-  );
-};
