+++ Frontend/features/dashboard/brands/BrandImageRender.tsx
@@ -1,11 +1,11 @@
-import { IMAGE_ASPECT_RATIO } from "@/constant/imageAspectRatio";
+import { IMAGE_ASPECT_RATIO, IMAGE_PLACEHOLDER } from "@/constant/image";
 import { BrandResponse } from "@/types";
 import Image from "next/image";
 
 export const BrandImageRender = (record: BrandResponse) => {
   return (
     <Image
-      src={record.imageUrl ?? "/image-not-found.jpg"}
+      src={record.imageUrl ?? IMAGE_PLACEHOLDER.DEFAULT}
       alt={record.name}
       width={300}
       height={200}
